+++ projects/utils/src/lib/functions/empty-to-dash/index.ts
@@ -1,0 +1,1 @@
+export * from './empty-to-dash.js';